chore(utxo-lib): move TxWeightCalculator to src
