+++ packages/utxo-lib/tests/txweight.test.ts
@@ -1,4 +1,4 @@
-import { TxWeightCalculator } from './compose.utils';
+import { TxWeightCalculator } from '../src/txWeightCalculator';
 
 // fixtures from:
 // https://github.com/trezor/trezor-firmware/blob/1fceca73da523c5bf2bb0f398c91e00c728bdbe0/core/tests/test_apps.bitcoin.txweight.py
